Sastasha: use int RNG
(cherry picked from commit e44f2be26ab96ab866d35d4ed849564e23dfef6f)

+++ src/scripts/instances/dungeons/Sastasha.cpp
@@ -59,7 +59,7 @@ public:
   {
     // Random coral
     auto& RNGMgr = Common::Service< World::Manager::RNGMgr >::ref();
-    instance.setCustomVar( Coral, static_cast< uint64_t >( RNGMgr.getRandGenerator< float >( Corals::Blue, Corals::Green ).next() ) );
+    instance.setCustomVar( Coral, RNGMgr.getRandGenerator< uint32_t >( Corals::Blue, Corals::Green ).next() );
 
     switch( instance.getCustomVar( Coral ) )
     {
